Move trackEvent function to /actions/index.js
